+++ spec/features/users_can_view_projects_spec.rb
@@ -78,6 +78,14 @@ RSpec.feature "Users can view a list of projects" do
       expect(page.find("ul.projects-list > li:nth-of-type(2)")).to have_content("100002")
       expect(page.find("ul.projects-list > li:nth-of-type(3)")).to have_content("100001")
     end
+
+    scenario "projects are grouped under conversion date headings" do
+      three_years_time = Date.today.beginning_of_month + 3.years
+
+      visit projects_path
+
+      expect(page).to have_css("h2", text: three_years_time.strftime("%B %Y openers"))
+    end
   end
 
   context "when the user is a regional delivery officer" do
@@ -132,7 +140,6 @@ RSpec.feature "Users can view a list of projects" do
 
     within urn.ancestor("li") do
       expect(page).to have_content("School type: #{project.establishment.type}")
-      expect(page).to have_content("Converting on: #{project.provisional_conversion_date.to_formatted_s(:govuk)}")
       expect(page).to have_content("Incoming trust: #{project.incoming_trust.name}")
       expect(page).to have_content("Local authority: #{project.establishment.local_authority}")
     end